requirements.txt
Django==1.8.3
Pillow==2.8.2
django-ckeditor==4.4.8
django-static-precompiler==0.9
easy-thumbnails==2.2
mysqlclient==1.3.6
requests==2.7.0
django-mptt==0.7.4
django-mptt-admin==0.2.1
djangorestframework==3.1.3
django-debug-toolbar==1.3.2
